Fixed bugs

+++ zoom_in.py
@@ -51,7 +51,7 @@ def do_final(inv, ext, speed):
     extlist = [ext]
     speedlist = [speed]
     num_repeats = 20
-    session_duration = sp['move_interval'] * (num_repeats) + 10
+    session_duration = int(sp['move_interval'] * (num_repeats) + 10)
     retdict = {
         'init_y': inv,
         'init_z': ext,
